requirements.txt
altgraph==0.17.4
asgiref==3.8.1
boto3==1.35.70
botocore==1.35.70
chardet==5.2.0
Django==4.1
django-cors-headers==4.2.0
django-filter==23.2
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
mysqlclient==2.2.6
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32-ctypes==0.2.3
reportlab==4.2.5
s3transfer==0.10.4
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2